VirtualScreen pixel width is now float

+++ include/VirtualScreen.h
@@ -9,14 +9,14 @@ namespace sn
     class VirtualScreen : public sf::Drawable
     {
     public:
-        void create (uint width, uint height, uint pixel_size, sf::Color default_color);
+        void create (uint width, uint height, float pixel_size, sf::Color default_color);
         void setPixel (std::size_t x, std::size_t y, sf::Color color);
 
     private:
         void draw(sf::RenderTarget& target, sf::RenderStates states) const;
 
         sf::Vector2<uint> m_screenSize;
-        uint m_pixelSize; //virtual pixel size in real pixels
+        float m_pixelSize; //virtual pixel size in real pixels
         sf::VertexArray m_vertices;
     };
 };
